MAINT: move sctypes from _detail to the main torch namespace
sctypes are user visible, so they needed to be exported to the
torch_np namespace anyway. Using them in the implementation
turns out to be a bad idea, because implementations in _detail
operate on the torch level. This way, looking up defaults or casting
had to go from torch.dtype to sctypes to torch.dtype. Which is weird
and inconvenient and plain extraneous.
