+++ torch_np/__init__.py
@@ -1,7 +1,8 @@
 from ._wrapper import *  # isort: skip  # XXX: currently this prevents circular imports
 from . import random
 from ._binary_ufuncs import *
-from ._detail._scalar_types import *
+# from ._detail._scalar_types import *
+from ._dtypes import *
 from ._detail._util import AxisError, UFuncTypeError
 from ._dtypes import *
 from ._getlimits import errstate, finfo, iinfo
